move number trait to lib
