+++ src/lib.rs
@@ -1,3 +1,26 @@
 pub mod linear_algebra;
 pub mod probability;
 pub mod set_theory;
+use std::ops::{Add, Div, Mul, Sub};
+
+pub trait Number:
+    Copy
+    + PartialOrd
+    + Ord
+    + Add<Output = Self>
+    + Sub<Output = Self>
+    + Mul<Output = Self>
+    + Div<Output = Self>
+{
+}
+
+impl<T> Number for T where
+    T: Copy
+        + PartialOrd
+        + Ord
+        + Add<Output = T>
+        + Sub<Output = T>
+        + Mul<Output = T>
+        + Div<Output = T>
+{
+}
